+++ libcxx/test/libcxx/mem/mem.res/nodiscard.verify.cpp
@@ -6,20 +6,66 @@
 //
 //===----------------------------------------------------------------------===//
 
-// UNSUPPORTED: c++03, c++11, c++14
+// REQUIRES: std-at-least-c++17
 
-// check that functions are marked [[nodiscard]] as an extension in C++17
-
-// [[nodiscard]] std::pmr::memory_resource::allocate(size_t, size_t);
-// [[nodiscard]] std::pmr::polymorphic_allocator<T>::allocate(size_t, size_t);
+// check that <memory_resource> functions are marked [[nodiscard]]
 
 #include <memory_resource>
 
-void f() {
-  std::pmr::memory_resource* res = nullptr;
-  res->allocate(0);    // expected-warning {{ignoring return value of function}}
-  res->allocate(0, 1); // expected-warning {{ignoring return value of function}}
+#include "test_macros.h"
 
-  std::pmr::polymorphic_allocator<int> poly;
-  poly.allocate(0); // expected-warning {{ignoring return value of function}}
+void test() {
+  {
+    std::pmr::memory_resource* r = std::pmr::null_memory_resource();
+
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r->allocate(1);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r->allocate(1, 1);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r->is_equal(*r);
+
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    std::pmr::get_default_resource();
+    // expected-warning@+1 {{ignoring return value of function declared with const attribute}}
+    std::pmr::new_delete_resource();
+    // expected-warning@+1 {{ignoring return value of function declared with const attribute}}
+    std::pmr::null_memory_resource();
+  }
+
+  {
+    std::pmr::monotonic_buffer_resource r;
+
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r.upstream_resource();
+  }
+
+  {
+    std::pmr::polymorphic_allocator<int> a;
+
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    a.allocate(1);
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    a.select_on_container_copy_construction();
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    a.resource();
+  }
+
+  {
+    std::pmr::synchronized_pool_resource r;
+
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r.upstream_resource();
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r.options();
+  }
+
+  {
+    std::pmr::unsynchronized_pool_resource r;
+
+    // expected-warning@+1 {{ignoring return value of function declared with 'nodiscard' attribute}}
+    r.upstream_resource();
+    // expected-warning@+1 {{ignoring return value of function declared with pure attribute}}
+    r.options();
+  }
 }
